+++ src/components/ArtworkUploader.tsx
@@ -4,7 +4,7 @@ import { css } from "g-style";
 
 import API from "services/memoir";
 
-import FilePicker from "components/form/FilePicker";
+import FilePicker from "components/molecules/form/FilePicker";
 
 const className = css({
   boxSizing: "border-box",
