+++ src/stories/components/molecules/form/FilePicker.stories.tsx
@@ -1,11 +1,11 @@
 import { h } from "preact";
 import { Meta, Story } from "@storybook/preact";
 
-import FilePicker, { Props } from "components/form/FilePicker";
+import FilePicker, { Props } from "components/molecules/form/FilePicker";
 
 const meta: Meta = {
   component: FilePicker,
-  title: "Components/Form/FilePicker",
+  title: "Components/Molecules/Form/FilePicker",
   argTypes: { onSelect: { action: "select" } },
 };
 export default meta;
