+++ src/stories/components/molecules/form/Submit.stories.tsx
@@ -1,11 +1,11 @@
 import { h } from "preact";
 import { Meta, Story } from "@storybook/preact";
 
-import Submit, { Props } from "components/form/Submit";
+import Submit, { Props } from "components/molecules/form/Submit";
 
 const meta: Meta = {
   component: Submit,
-  title: "Components/Form/Submit",
+  title: "Components/Molecules/Form/Submit",
   argTypes: { onClick: { action: "clicked" } },
 };
 export default meta;
